fix(sync): address lock review findings
