+++ os/arceos/modules/axsync/src/axtest.rs
@@ -2,35 +2,35 @@ use axtest::prelude::*;
 
 #[axtest]
 fn spin_rwlock_constants_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_constants_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_constants_hold_for_test());
 }
 
 #[axtest]
 fn spin_rwlock_state_logic_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_state_logic_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_state_logic_hold_for_test());
 }
 
 #[axtest]
 fn spin_rwlock_constants_and_phantom_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_constants_and_phantom_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_constants_and_phantom_hold_for_test());
 }
 
 #[axtest]
 fn spin_rwlock_state_transitions_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_state_transitions_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_state_transitions_hold_for_test());
 }
 
 #[axtest]
 fn spin_rwlock_guard_types_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_guard_types_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_guard_types_hold_for_test());
 }
 
 #[axtest]
 fn spin_rwlock_lockdep_and_feature_config_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_lockdep_and_feature_config_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_lockdep_and_feature_config_hold_for_test());
 }
 
 #[axtest]
 fn spin_rwlock_reader_writer_state_combinations_hold() {
-    ax_assert!(crate::spin_rwlock::rwlock_reader_writer_state_combinations_hold_for_test());
+    ax_assert!(crate::spin::rwlock::rwlock_reader_writer_state_combinations_hold_for_test());
 }
