+++ virtualization/axvm/src/arch/x86_64/mod.rs
@@ -14,7 +14,7 @@ use std::{
     time::Duration,
 };
 
-use ax_std::os::arceos::sync::{RawSpinLock, RawSpinLockGuard};
+use ax_std::os::arceos::sync::{IrqSafeMutex, IrqSafeMutexGuard};
 use axdevice::*;
 use axdevice_base::*;
 use axvm_types::{VmBackendError as BackendError, VmBackendResult as BackendResult, *};
@@ -505,9 +505,9 @@ struct X86IoApicModel {
 /// VM-owned domain.
 pub(super) struct X86InterruptDomain {
     wired: Arc<X86WiredState>,
-    inputs: RawSpinLock<BTreeMap<usize, (InterruptTriggerMode, WiredIrqInput)>>,
-    forwarding: RawSpinLock<irq::X86IoApicForwardingState>,
-    forwarding_hooks: RawSpinLock<std::vec::Vec<host_irq::IrqHandle>>,
+    inputs: IrqSafeMutex<BTreeMap<usize, (InterruptTriggerMode, WiredIrqInput)>>,
+    forwarding: IrqSafeMutex<irq::X86IoApicForwardingState>,
+    forwarding_hooks: IrqSafeMutex<std::vec::Vec<host_irq::IrqHandle>>,
 }
 
 struct X86WiredState {
@@ -534,22 +534,16 @@ impl ServiceKey for X86InterruptDomainRuntimeKey {
 impl X86InterruptDomain {
     fn inputs(
         &self,
-    ) -> RawSpinLockGuard<'_, BTreeMap<usize, (InterruptTriggerMode, WiredIrqInput)>> {
-        // SAFETY: x86 interrupt-domain callers already run with local
-        // preemption/IRQ re-entry excluded by the guest-entry or IRQ path.
-        unsafe { self.inputs.lock_raw() }
+    ) -> IrqSafeMutexGuard<'_, BTreeMap<usize, (InterruptTriggerMode, WiredIrqInput)>> {
+        self.inputs.lock()
     }
 
-    fn forwarding(&self) -> RawSpinLockGuard<'_, irq::X86IoApicForwardingState> {
-        // SAFETY: forwarding state is accessed only from guest-entry and host
-        // IRQ paths that exclude same-CPU re-entry.
-        unsafe { self.forwarding.lock_raw() }
+    fn forwarding(&self) -> IrqSafeMutexGuard<'_, irq::X86IoApicForwardingState> {
+        self.forwarding.lock()
     }
 
-    fn forwarding_hooks(&self) -> RawSpinLockGuard<'_, std::vec::Vec<host_irq::IrqHandle>> {
-        // SAFETY: hook publication and teardown are serialized by the VM
-        // lifecycle before the raw lock is acquired.
-        unsafe { self.forwarding_hooks.lock_raw() }
+    fn forwarding_hooks(&self) -> IrqSafeMutexGuard<'_, std::vec::Vec<host_irq::IrqHandle>> {
+        self.forwarding_hooks.lock()
     }
 
     fn new(vm_id: usize, ioapic: Arc<dyn X86IoApicDeviceOps>) -> Self {
@@ -560,9 +554,9 @@ impl X86InterruptDomain {
                 pending_level: AtomicUsize::new(0),
                 kick: DeferredVcpuKick::new(vm_id),
             }),
-            inputs: RawSpinLock::new(BTreeMap::new()),
-            forwarding: RawSpinLock::new(irq::X86IoApicForwardingState::new()),
-            forwarding_hooks: RawSpinLock::new(std::vec::Vec::new()),
+            inputs: IrqSafeMutex::new(BTreeMap::new()),
+            forwarding: IrqSafeMutex::new(irq::X86IoApicForwardingState::new()),
+            forwarding_hooks: IrqSafeMutex::new(std::vec::Vec::new()),
         }
     }
 
@@ -943,12 +937,28 @@ fn restore_host_interrupt_flag(host_rflags: u64) {
 
 #[cfg(test)]
 mod tests {
+    use ax_std::os::arceos::sync::IrqSafeMutex;
+
     use super::*;
+
     fn assert_x86_exit_type<T: VmArchVcpuOps<Exit = X86VmExit>>() {}
+
+    fn assert_irq_safe_lock<T: ?Sized>(_: &IrqSafeMutex<T>) {}
 
     #[test]
     fn axvm_x86_vcpu_uses_x86_exit_type() {
         assert_x86_exit_type::<AxvmX86Vcpu>();
+    }
+
+    #[test]
+    fn interrupt_domain_shared_state_uses_irq_safe_locks() {
+        fn check(domain: &X86InterruptDomain) {
+            assert_irq_safe_lock(&domain.inputs);
+            assert_irq_safe_lock(&domain.forwarding);
+            assert_irq_safe_lock(&domain.forwarding_hooks);
+        }
+
+        let _ = check as fn(&X86InterruptDomain);
     }
 
     #[test]
